Added font. Refactored every class into its own file. Added + and - lines outputhandling
